fix[upa]: clean dupSubmissionIdx unpacking. move and rename packing functions.
